Use resource_id for quarantine resourcedata files
Due to a bug, we encountered a situation where only a part of the
resource_id was found in the acquired evidence.
This led to the invalid assumption on my end that sometimes
defender will only use a part of the resource id for the filename of
a ResourceData file.

Now, we assume the a quarantine entry's resource_id will be the
filename found in the ResourceData folder.
This also aligns with what we saw when investigating mpengine.dll.