Implement Citrix filesystem structure using create
The previous approach returned a virtualfilesystem on detect().
As I understand it, an OS plugin should return the system volume on
detect, and bootstrap the filesystem in the create classmethod.
Also added a check for /log to distinguish ramdisk from system volume.
Lastly, account for usage on *only* the ramdisk by falling back to the
ramdisk as system volume if the harddisk is not available.