Change the RunKeyRecord path field to string
The "path" value of a run key is actually a full command. If this
command does not have just a path but also things like command
parameters, this can not be parsed as a path type.

The field is now treated as a normal string to prevent parse errors and
in anticipation of a proper command field type.

(DIS-2563)