Fix permission issues when tests are run on windows
On Windows, files can not be opened when already open. This gives issues
when using temporary files. By not deleting a temporaryfile after it is
closed and closing it after it is created, this is circumvented.

By moving all these files into pytest's tmp_path, the tmp_path mechanism
will take care of removing these temporary files.