Fix NameSpacePlugin function filtering
The previous check for subplugin export functions was too strict.
It did not account for subplugins reusing export methods of others.
Also improved handling unsupported plugins.
Lastly, added a check for when a subplugin has an export function with
the same name as the NameSpacePlugin to which it belongs.
Added tests for the scenarios outlined above.