Improve driveletter selection for windows disks
As of c97c344 filesystems are always mounted. In the case of "fake" NTFS
filesystems, as used by e.g. the dir and tar loaders, this means such a
filesystem can be present under up to three different mount points,
e.g.: sysvol, C: and $fs$\fs0

When a plugin does not use the root filesystem but the bare ntfs
filesystem to parse ntfs specific properties, e.g. the mft plugin, it
extends paths with a "driveletter" (read: mountpoint) to construct a
path that is reachable from the root filesystem.

The old setup always returned the last mount point a drive was mounted
under. The new setup changes this to return the first mount point present
in the following order:

1. the first mount point with a letter and colon, e.g. C:,
2. the first mount point not starting with $fs$, e.g. sysvol,
3. the first mount point starting with $fs$, e.g. $fs$\fs0.