Add support for Fortinet AES encrypted firmware images
This PR also adds the already dumped AES keys for convienience.
It contains dumped FortiGate keys up to the latest v7.6.1.

The AES keys are extracted from the 256 byte rootfs.gz footer which is
encrypted using RSA. The RSA public key needed to decrypt this is
encrypted using ChaCha20, both are stored in the flatkc kernel file.

Closes #968