Ensure shell field is of type posix_path in UnixUserRecord
Ensure shell field is of type posix_path in UnixUserRecord and update tests accordingly. Apparently, in dissect, the test seemed to fail only on windows, so i made sure that the type is a posix_path, just like the home attribute. If anything needs to change, just let me know in the PR.